+++ changelog.md
@@ -1,3 +1,7 @@
+1.6.0
+- Prevented stacking of ender bag item
+- Allowed renaming the ender bag at the anvil
+
 1.5.0
 - Fixed interact event throwing ender bag in some cases (fences, jukebox, cauldron)
 - Fixed ender bag being usable as its base item in crafting